INTELIGENCIA ARTIFICIAL/requirements.txt
Flask
Flask-Cors
requests
duckduckgo_search
python-dotenv
PyPDF2
python-docx
Pillow
beautifulsoup4
huggingface-hub
transformers
torch
torchvision
torchaudio
accelerate
sentence-transformers
numpy
scikit-learn
pandas
gunicorn
lxml
soupsieve
tokenizers
tqdm
filelock
fsspec
packaging
pyyaml
regex
safetensors
typing-extensions
charset-normalizer
idna
urllib3
certifi
# Novos para OCR e PDF avançado
pdfplumber
pytesseract
